requirements.txt
streamlit==1.39
openai==1.51.1
newsapi-python==0.2.7